first comment

+++ database/migrations/2019_09_11_150205_create_order_items_table.php
@@ -18,7 +18,7 @@ class CreateOrderItemsTable extends Migration
 
             $table->unsignedBigInteger('order_id')->index();
             $table->unsignedBigInteger('product_id')->index();
-            $table->unsignedInteger('quantity');
+            $table->unsignedBigInteger('quantity');
             $table->decimal('price', 20, 6);
 
             $table->foreign('order_id')->references('id')->on('orders')->onDelete('cascade');
